More robust links and images.
Added the pubPath utility function to get the path for the public
directory using basePath variable from the next.config.js file.
Switched to <Link> NextJS component which substitutes the base path
automatically. Also changed some mockup data (to be removed in the
future).

These changes are mostly needed to support deployment on Github pages,
especially as a subproject for which the base path is
user.github.io/reponame.