Changed "Festival" to server component.
This is an attempt to address a bug that prevents slideshow in the
"Festival" page, unless it is refreshed. Because the same components
works well in the Home page and the only difference is that Home is a
server component, we try solve the bug by switching to a server
component even in the Festival page.