Menu page updated with data loading.
The "Menu" page now loads the list of sub-menus from a JSON file (this
will be localized in the future). The page displays a loading marker
while files are loaded, but doesn't give a feedback in case of error
(to be implemented).

The Accordion used within the menu has been slightly changed to use
the NextJS <Image> component and to address the case of missing or
wrong image. Currently the missing image is replaced by a generic svg,
but in the future this could be changed based on the sub menus.