Added subcategories to the menu.
For those categories that have a further sub categorization, we added
the corresponding hierarchy in the displayed menu. Accordingly, the
JSON formate used internally (parsed from the online csv), is changed.

We also changed the red shade of the Laguna Blues logo, even though such
a shade may be too dark (see previous comments on that).