Added category notes
We added a special behavior on the menu for items with number 900 or
more. When one of this is encountered, the displayed item is a InfoItem
rather then MenuItem.

This can be used to add information at the end of category since the
item are sorted numerically and these will supposedly appear last. If
multiple such items are present in the data file, they are displayed in
the order of number given.